Added unit tests and split up the two classes to get this package to look like a 21st century PHP library :-p
