+++ .travis.yml
@@ -1,0 +1,21 @@
+language: php
+php:
+  - '5.6'
+  - '7.0'
+  - '7.1'
+  - '7.2'
+  - hhvm
+  - nightly
+
+matrix:
+  allow_failures:
+    - php: 7.2
+    - php: hhvm
+    - php: nightly
+
+before_script:
+  - composer self-update
+  - composer install
+
+script:
+  - vendor/bin/phpunit
