+++ composer.json
@@ -18,6 +18,9 @@
         "php": ">=5.3.3"
     },
     "autoload": {
-        "psr-0": { "PWGen": "" }
+        "psr-4": { "PWGen\\": "src/" }
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^6.4"
     }
 }
