+++ phpunit.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0"?>
+<phpunit bootstrap="tests/bootstrap.php"
+         colors="true"
+         syntaxCheck="true"
+>
+
+    <testsuites>
+        <testsuite name="PWGen password generator">
+            <directory>./tests/PWGen</directory>
+        </testsuite>
+    </testsuites>
+
+</phpunit>
