+++ tests/bootstrap.php
@@ -1,0 +1,5 @@
+<?php
+
+print 'PHP v' . phpversion() . "\n";
+
+require_once __DIR__ . "/../vendor/autoload.php";
